pyproject.toml
[project]
name = "lg-playground"
version = "0.1.0"
description = "playground for LangGraph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "langchain[openai]>=0.3.25",
    "langgraph>=0.4.1",
    "langgraph-cli[inmem]>=0.2.8",
    "langgraph-sdk>=0.1.66",
    "langsmith>=0.3.42",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "elasticsearch>=8.18.0,<9.0.0",
    "pytest>=8.3.5",
    "coverage>=7.8.0",
    "rich>=14.0.0",
]

[project.optional-dependencies]
lint = [
    "black>=25.1.0",
    "ruff>=0.11.8",
]

test = [
    "pytest>=8.3.5",
    "coverage>=7.8.0",
    "pytest-mock>=3.14.0",
]

dev = [
    "ipython>=8.22.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
